+++ docs/changelog/changelog.md
@@ -4,7 +4,9 @@
 
 ### Features
 
-- New `Report` App added.
+- New `Report` App added for easily creating printable reports.
+  - Located at `django_spire.metric.report` with its own auth controller.
+  - Check out the documentation for more information.
 
 
 ## v0.23.12